new file:   API/KAKAO_POI.js 	renamed:    POI.js -> API/POI.js 	new file:   API/Ultra_Forecast.js 	renamed:    cal_time_date.js -> API/cal_time_date.js 	renamed:    public_transport.js -> API/public_transport.js 	deleted:    Ultra_Forecast.js 	modified:   main.js 	modified:   pposongtime.js 	modified:   public/css/mainFunc-style.css 	new file:   public/js/convert_XY.js 	modified:   public/js/mainFunc.js 	modified:   views/mainFunc.html
