+++ public/css/mainFunc-style.css
@@ -263,9 +263,18 @@ hr {
 }
 
 .h {
-	font-size: 50px;
+	font-size: 40px;
 	background-color: #add6ec;
 	color: white;
-	width: 400px;
+	width: 90%;
 	margin: 2px;
+}
+
+.autocomplete-list {
+	max-height: 200px;
+	/* 원하는 최대 높이 설정 */
+	overflow-y: auto;
+	/* 세로 스크롤 적용 */
+	border: 1px solid #ccc;
+	/* 스크롤이 보이도록 테두리 설정 (선택 사항) */
 }
